Fix the issue with python and fish
The problem seen earlier is that miniconda ships with a binary named
clear which quite possibly does the same thing. However when you don't
have miniconda selected as the global python it fails to execute.

Instead of setting miniconda as the global python it was decided to
remove it and determine a decent way of managing python in the future.