nvim: Install neovim on Ubuntu 21.10
The directory installation was incorrectly done and would create a
folder named "~/.config/nvim/{backups,swaps,undo}" instead of expanding
and creating several directories. Instead of trading speed let's do it
the correct way in ansible.

When running neovim wihtout any plugins installed it would fail. This
caused the plugins to not be installed. Fix this issue by only loading
the colorscheme if the plugin is sourced for it.